Removed stable_networking.
This was introduced by duke3d_w32 before xDuke3D, and supposedly only supports up to 2 players. I have found no trace of its use, even in modern matchmaking software.

Since we want to keep the engine small and understandable, I decided to remove it (it also removes the enet dependency). More cleanup to be done in this area.